Add dynamic waitlist URL throughout website
This update imports a function to generate the waitlist URL, and applies it in several components replacing previously hard-coded waitlist links. By using a function, the waitlist URL can now be retrieved dynamically, resulting in better code maintainability and consistency. In addition, react-helmet package was added to the dependencies in package.json.